backend/requirements.txt
pypdf2
toml
langchain
transformers
torch
python-dotenv
pinecone-client
google-generativeai
"fastapi[standard]"